Address PR review comments
1. Proto: Remove unpopulated SourceIngestionInfo message and repeated field;
   reserve field number 6 for future use
2. LeaderProducerCallback: Add clarifying comment that callback latency
   measurement matches the existing recordProducerCallBackLatency pattern;
   remove unnecessary cast
3. VeniceIngestionMonitorServiceImpl: Add close() method (implements
Closeable)
   that shuts down the scheduler and cleans up all active sessions; fix race
   condition by using putIfAbsent for session registration before attaching
   monitor to PCS
4. AdminTool: Use GrpcUtils.buildChannelCredentials(sslFactory) for proper
   SSL channel setup instead of bare useTransportSecurity()

+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/kafka/consumer/LeaderProducerCallback.java
@@ -228,8 +228,10 @@ public class LeaderProducerCallback implements ChunkAwareCallback {
         }
         PartitionIngestionMonitor callbackMonitor = partitionConsumptionState.getIngestionMonitor();
         if (callbackMonitor != null) {
-          callbackMonitor.recordLeaderCallbackLatencyNs(
-              (long) ((System.currentTimeMillis() - currentTimeForMetricsMs) * 1_000_000L));
+          // Measures time spent in callback processing (drainer queuing, chunk handling, etc.)
+          // Same measurement as recordProducerCallBackLatency above, converted to ns for precision
+          callbackMonitor
+              .recordLeaderCallbackLatencyNs((System.currentTimeMillis() - currentTimeForMetricsMs) * 1_000_000L);
           callbackMonitor.recordLeaderProduced(producedRecordSize);
         }
         this.onCompletionCallback.accept(produceResult);
